+++ packages/jam-pvm/src/instructions/ixs/one_reg_two_imm_ixs.ts
@@ -141,8 +141,8 @@ if (import.meta.vitest) {
           context,
           store_imm_ind_u8,
           10,
-          0x10,
-          0x12,
+          0x10 as u32,
+          0x12 as u32,
         );
         expect((p_context.memory.setBytes as Mock).mock.calls).toEqual([
           [0x1010, new Uint8Array([0x12])],
@@ -156,8 +156,8 @@ if (import.meta.vitest) {
           context,
           store_imm_ind_u16,
           10,
-          0x10,
-          0x1234,
+          0x10 as u32,
+          0x1234 as u32,
         );
         expect((p_context.memory.setBytes as Mock).mock.calls).toEqual([
           [0x1010, new Uint8Array([0x34, 0x12])],
@@ -171,8 +171,8 @@ if (import.meta.vitest) {
           context,
           store_imm_ind_u32,
           10,
-          0x10,
-          0x12345678,
+          0x10 as u32,
+          0x12345678 as u32,
         );
         expect((p_context.memory.setBytes as Mock).mock.calls).toEqual([
           [0x1010, new Uint8Array([0x78, 0x56, 0x34, 0x12])],
